+++ src/__tests__/fixtures/EditionCardFixture.ts
@@ -106,3 +106,35 @@ export const eddEdition: WorkEdition = {
     },
   ],
 };
+
+export const upEdition: WorkEdition = {
+  ...fullEdition,
+  items: [
+    {
+      links: [
+        {
+          url: "test-link-url",
+          link_id: 12,
+          mediaType: "application/epub+xml",
+          flags: {
+            catalog: false,
+            download: false,
+            reader: true,
+            nypl_login: true,
+          },
+        },
+        {
+          url: "test-link-url-2",
+          link_id: 23,
+          mediaType: "application/epub+zip",
+          flags: {
+            catalog: false,
+            download: true,
+            reader: false,
+            nypl_login: true,
+          },
+        },
+      ],
+    },
+  ],
+};