+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add Physical Edition badge and Scan and Deliver blurb to EDD editions
 - Add Library Card Required and Featured Editon badges
 - Update styles for work and edition pages
+- Add login buttons and info blurb UP items
 
 ## [0.17.5]